refactor(version-compare): extract pure isNewer so tests skip update-check chain
test/update-check.test.ts only tests version comparison, but it was
importing server/update-check.ts which drags in ws.ts + pipeline.ts +
db.ts — hundreds of ms of module-load cost and open handles that
hung the test runner under concurrency.

Moves isNewer to server/version-compare.ts (pure, no side imports).
update-check.ts re-exports it for production callers. Test now imports
from version-compare directly — 14ms for the whole file.
